removed custom http library, refactored weatherbit
I decided to just entirely remove the custom http library since it was
really just a needless abstraction over axios. The error handling code
that it contained was overly complex and arguably not needed at all.

Now the weatherbit library directly uses axios and returns the axios
promise to the calling code, which awaits on it.

It would be ideal to have a global axios config that all code can use,
but this won't be necessary until something more than the weatherbit
code uses axios.